pyproject.toml
[project]
name = "gitlab-agent"
version = "0.0.1"
description = "LLM agent to process gitlab web hooks"
authors = [
    {name = "Anatoliy A Aksenov",email = "anatoliy@aksenov.org"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<=1.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langgraph (>=0.6.5,<0.7.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "langchain-mcp-adapters (>=0.1.9,<0.2.0)",
    "retry (>=0.9.2,<0.10.0)",
    "langchain-openai (>=0.3.32,<0.4.0)",
    "python-gitlab (>=6.2.0,<7.0.0)",
    "langgraph-checkpoint-postgres (>=2.0.23,<3.0.0)",
    "psycopg[binary,pool] (>=3.2.10,<4.0.0)",
    "langfuse (>=3.5.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
